Move login pages to src/pages/Login
For consistency with other entity pages, move src/login/*.tsx into
src/pages/Login/ and update the corresponding imports in AppRoutes.tsx.

Fixes #722
