+++ src/navigation/AppRoutes.tsx
@@ -42,15 +42,15 @@ import KrbTicketPolicy from "src/pages/KrbTicketPolicy/KrbTicketPolicy";
 import UserGroupsTabs from "src/pages/UserGroups/UserGroupsTabs";
 import HostGroupsTabs from "src/pages/HostGroups/HostGroupsTabs";
 import IDViewsTabs from "src/pages/IDViews/IDViewsTabs";
-import LoginMainPage from "src/login/LoginMainPage";
+import LoginMainPage from "src/pages/Login/LoginMainPage";
 import NetgroupsTabs from "src/pages/Netgroups/NetgroupsTabs";
 import HBACServicesTabs from "src/pages/HBACServices/HBACServicesTabs";
 import HBACRulesTabs from "src/pages/HBACRules/HBACRulesTabs";
 import HBACServiceGroupsTabs from "src/pages/HBACServiceGroups/HBACServiceGroupsTabs";
-import ResetPasswordPage from "src/login/ResetPasswordPage";
+import ResetPasswordPage from "src/pages/Login/ResetPasswordPage";
 import SetupBrowserConfig from "src/pages/SetupBrowserConfig";
 import Configuration from "src/pages/Configuration/Configuration";
-import SyncOtpPage from "src/login/SyncOtpPage";
+import SyncOtpPage from "src/pages/Login/SyncOtpPage";
 import SubordinateIDs from "src/pages/SubordinateIDs/SubordinateIDs";
 import SubIdsStatistics from "src/pages/SubordinateIDs/SubIdsStatistics";
 import SubIdsTabs from "src/pages/SubordinateIDs/SubIdsTabs";
